+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Loader: melonJS now dynamically assign parser/preload functions to known asset types
 - Loader: add the possibility to specify a user defined parser/preload function for a given asset type
 - Renderer: the default renderer instance is now passed as parameter when emitting the global `VIDEO_INIT` event
+- Renderer: add path like methods (`beginPath`, `lineTo`, `moveTo`, `rect`, `roundRect`, `closePath`, `fill` and `stroke`)
 
 ### Changed
 - Geometry: optimize Path2d triangulation by only recalulcating all triangless when the path is modified